pyproject.toml
[tool.poetry]
name = "auto-invoicer"
version = "0.1.0"
description = ""
authors = ["Christopher Carroll Smith <75859865+chriscarrollsmith@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "auto_invoicer"}]

[tool.poetry.dependencies]
python = "^3.9"
datetime = "^5.1"
python-dateutil = "^2.8.2"
markdown = "^3.4.3"
python-dotenv = "^1.0.0"
pyhtml2pdf = "^0.0.6"
python-crontab = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
